quadratic-functions/requirements.txt
flask>=2.3.0
google-generativeai>=0.3.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.17.0
python-dotenv>=1.0.0 